Share startup session caching across local and remote agent hosts
Hoist the session-summary persistence layer from RemoteAgentHostSessionsProvider
into BaseAgentHostSessionsProvider so the local agent host also shows its
sessions immediately on Agents-window startup, before the in-process agent host
has started and the first listSessions() round-trip returns.

- Base now owns serialize/deserialize, _metaByRawId/_cacheDirty bookkeeping, and
  _loadCachedSessions()/_persistCache(), gated behind an explicit
  _enableSessionCachePersistence(storageKey) opt-in that subclasses call at the
  end of their constructor (once identity fields needed by createAdapter are set).
- A _shouldTrackSessionCacheChanges() hook preserves the remote provider's
  offline (_unpublished) gate.
- Remote provider drops the migrated code and just supplies its per-authority
  key; local provider supplies the fixed key localAgentHost.cachedSessions.
